requirements.txt
transitions
openai
tiktoken
python-dotenv
pyflakes
psycopg2-binary
sqlalchemy
tqdm
console-menu
requests
bcrypt